refactor: useSearchPage の冗長なハンドラ関数を簡素化
- handleQueryChange と handleHistorySearch を統合（同一ロジック）
- handleSelect を削除し selectItem を直接 return
- handleThemeHistorySelect を削除し setTheme を直接利用
- mainStyle をコンポーネント側（SearchPage.tsx）に移動
- useMemo, SearchResultItem の不要な import を削除

Closes #202
